Player class uploaded and completed
Player class includes basic information needed from each player. A constructor initializes the instance variables. There are methods to set and get the player name, player ID, and player balance. There is a method for setting a random player ID. There are also methods for the player to deposit and withdraw money during games when needed.