Update Craps.java
Imported Scanner to use the Scanner class in this file. Added "int" to the moneyEarned variable for it to be declared and initialized. Changed exception to built-in exception IllegalArgumentException so that we don't have to create a custom exception for it to work. Overall, fixed errors in the file so it is runnable.